+++ lib/edge/src/edge_shard/snapshots.rs
