+++ lib/edge/src/read_view/count.rs
