+++ lib/edge/src/read_view/facet.rs
