+++ lib/edge/src/read_view/info.rs
