+++ lib/edge/src/read_view/retrieve.rs
